scala type system <3
